requirements.txt
aiogram~=2.11.2
aiohttp==3.7.3
aioredis==1.3.1
python-dotenv~=0.15.0
Pillow~=8.0.1
sql~=0.4.0
asyncpg~=0.21.0

numpy~=1.19.4
torch~=1.7.0
torchvision~=0.8.2
opencv-contrib-python~=4.4.0.46
